refactor: simplify session launch plumbing
