+++ apps/desktop/src-tauri/crates/host-domain/src/runtime/registry/capabilities.rs
@@ -64,6 +64,61 @@ impl fmt::Display for RuntimeSessionStartMode {
     }
 }
 
+struct LaunchStartModeRequirement {
+    id: &'static str,
+    modes: &'static [RuntimeSessionStartMode],
+}
+
+const LAUNCH_START_MODE_REQUIREMENTS: &[LaunchStartModeRequirement] = &[
+    LaunchStartModeRequirement {
+        id: "spec_initial",
+        modes: &[RuntimeSessionStartMode::Fresh],
+    },
+    LaunchStartModeRequirement {
+        id: "planner_initial",
+        modes: &[RuntimeSessionStartMode::Fresh],
+    },
+    LaunchStartModeRequirement {
+        id: "build_implementation_start",
+        modes: &[RuntimeSessionStartMode::Fresh],
+    },
+    LaunchStartModeRequirement {
+        id: "build_after_qa_rejected",
+        modes: &[
+            RuntimeSessionStartMode::Fresh,
+            RuntimeSessionStartMode::Reuse,
+        ],
+    },
+    LaunchStartModeRequirement {
+        id: "build_after_human_request_changes",
+        modes: &[
+            RuntimeSessionStartMode::Fresh,
+            RuntimeSessionStartMode::Reuse,
+        ],
+    },
+    LaunchStartModeRequirement {
+        id: "build_pull_request_generation",
+        modes: &[
+            RuntimeSessionStartMode::Reuse,
+            RuntimeSessionStartMode::Fork,
+        ],
+    },
+    LaunchStartModeRequirement {
+        id: "build_rebase_conflict_resolution",
+        modes: &[
+            RuntimeSessionStartMode::Fresh,
+            RuntimeSessionStartMode::Reuse,
+        ],
+    },
+    LaunchStartModeRequirement {
+        id: "qa_review",
+        modes: &[
+            RuntimeSessionStartMode::Fresh,
+            RuntimeSessionStartMode::Reuse,
+        ],
+    },
+];
+
 #[derive(Debug, Clone, Copy, Serialize, Deserialize, PartialEq, Eq, Hash)]
 #[serde(rename_all = "snake_case")]
 pub enum RuntimeForkTarget {
@@ -619,25 +674,25 @@ impl RuntimeCapabilities {
     }
 
     pub(super) fn launch_config_errors(&self) -> Vec<String> {
-        let required_pull_request_modes = [
-            RuntimeSessionStartMode::Reuse,
-            RuntimeSessionStartMode::Fork,
-        ];
-        let missing_pull_request_modes = required_pull_request_modes
+        LAUNCH_START_MODE_REQUIREMENTS
             .iter()
-            .copied()
-            .filter(|mode| !self.session_lifecycle.supported_start_modes.contains(mode))
-            .collect::<Vec<_>>();
-        if missing_pull_request_modes.is_empty() {
-            return Vec::new();
-        }
-        vec![format!(
-            "[launch_scoped] launch action build_pull_request_generation requires start modes: {}",
-            missing_pull_request_modes
-                .into_iter()
-                .map(|mode| mode.to_string())
-                .collect::<Vec<_>>()
-                .join(", ")
-        )]
+            .filter_map(|requirement| {
+                let missing_modes = requirement
+                    .modes
+                    .iter()
+                    .copied()
+                    .filter(|mode| !self.session_lifecycle.supported_start_modes.contains(mode))
+                    .map(|mode| mode.to_string())
+                    .collect::<Vec<_>>();
+                if missing_modes.is_empty() {
+                    return None;
+                }
+                Some(format!(
+                    "[launch_scoped] launch action {} requires start modes: {}",
+                    requirement.id,
+                    missing_modes.join(", ")
+                ))
+            })
+            .collect()
     }
 }
