+++ packages/frontend/src/pages/agents/query-sync/use-agent-studio-query-session-sync.ts
@@ -50,8 +50,6 @@ export function useAgentStudioQuerySessionSync({
       [AGENT_STUDIO_QUERY_KEYS.task]: undefined,
       [AGENT_STUDIO_QUERY_KEYS.session]: undefined,
       [AGENT_STUDIO_QUERY_KEYS.agent]: undefined,
-      [AGENT_STUDIO_QUERY_KEYS.autostart]: undefined,
-      [AGENT_STUDIO_QUERY_KEYS.start]: undefined,
     });
   }, [
     isLoadingTasks,
