+++ packages/frontend/src/pages/agents/use-agent-studio-query-session-sync.test.tsx
@@ -63,8 +63,6 @@ describe("useAgentStudioQuerySessionSync", () => {
           task: undefined,
           session: undefined,
           agent: undefined,
-          autostart: undefined,
-          start: undefined,
         },
       ]);
     } finally {
